feat: moved lifecycles to decorator
